Update library for provider configurations
